requirements.txt
discord.py
pyaudio
speechrecognition
pyttsx3
pydub
asyncio
